Add AboutReligionView and navigation to it from About Religion section
This commit adds the AboutReligionView to the project and implements navigation to it from the About Religion section in the home page. The RowLabelWithShowMore widget is updated to include a Bounceable widget that triggers the navigation to the AboutReligionView when clicked. The onGenerateRoutes function is also updated to include a case for the AboutReligionView route.