Add video youtube entity, model and repo implementation
add about_religion_videos.json

This commit adds the VideoYoutubeEntity and VideoYoutubeModel classes to the
about_religion domain entities and data models directories respectively. It
also adds the VideoYoutubeRepoImpl class to the domain data repos directory.
The VideoYoutubeModel class extends the VideoYoutubeEntity class and has a
fromJson factory method to parse the JSON data. The VideoYoutubeRepoImpl class
loads and parses the JSON file from assets and returns a list of
VideoYoutubeEntity objects. Additionally, the pubspec.yaml file is updated to
include the assets/json/ directory.