Refactor AboutReligion feature to use ManageAboutReligionCubit
This commit refactors the AboutReligion feature to use the ManageAboutReligionCubit for state management. The AboutReligionViewBody and AboutReligionColumn widgets have been updated to use BlocBuilder to display the list of videos fetched from the API. The AboutReligionItem widget has also been updated to display the thumbnail and title of each video. This change improves the maintainability and scalability of the feature.