Add shortcut view and update UI
Added a new feature to display shortcuts on the home screen. Also, updated the UI of the About Religion view by adding a custom font style to the app bar title and reducing the padding of the search icon. Additionally, made some minor changes to the CustomShortCutImageShape widget and the ShortCutItemsListView widget. Finally, added a new view for the shortcuts and updated the onGenerateRoutes function to handle the new route.