Refactor code to implement favorite feature
This commit refactors the code to implement a favorite feature. It adds a `FavoriteEntity` abstract class and a `FavoriteRepository` class to handle storing and retrieving favorite entities. The `VideoYoutubeModel`, `AzkarItemEntity`, and `HadithEntity` classes now extend `FavoriteEntity` and implement the `toJson()` and `getIdentifier()` methods. The `FavoritesView` is also added to display the user's favorite items. Finally, the `getItSetup()` function is updated to register the `FavoriteRepository` as a singleton.