Add prayer background image and update UI in prayer times view header
This commit adds a new prayer background image and updates the UI in the prayer times view header. The new image is used as the background for the header and the UI for the buttons to find the nearest mosque and qiblah have been updated to use a container with a custom style instead of a circle avatar. The commit also includes an update to the `getIdentifier()` method in the `AzkarItemEntity` class to include additional fields.