Refactor hadith processing to use isolates for heavy tasks
This commit refactors the `HadithProcessingCubit` to use isolates for heavy tasks such as fetching sections and ahadith for a single book. This improves performance and prevents the UI from freezing. The `_fetchHadithForBook` method now spawns an isolate to process the heavy tasks and returns a list of `HadithEntity` objects. The `_isolateProcess` method is used to process the heavy tasks in the isolate. The `_getSectionNumber` method is now static and logs the section number for debugging purposes.