Refactor HadithProcessingCubit to load and save data from/to shared preferences
This commit refactors the HadithProcessingCubit to load data from shared preferences on creation and save data to shared preferences when closing the cubit. The hadiths are stored as JSON strings in shared preferences. This improves the performance of the app by reducing the time it takes to process the hadiths.