Add hourly Azkar notifications using Android Alarm Manager
This commit adds the functionality to show hourly Azkar notifications using Android Alarm Manager. The `_startHourlyAzkarNotifications()` method is called in the `initState()` of the `NavigationPage` widget to start the periodic notifications. The `periodicAzkarFetch()` method is the callback function that fetches a random Azkar and shows the notification using the `NotificationService`. The `fetchRandomAzkarWithLimit()` method in the `NotificationService` is made public to be used in the callback function. Additionally, the `AndroidAlarmManager` is initialized in the `main()` method.