feat: Add Android intent and notification channel to notifications service
This commit adds the `android_intent_plus` and `android_notification_channel` packages to the `NotificationService` class in order to request notification permissions and auto-start permission. It also updates the `showAzkarNotification` and `scheduleAzkarNotification` methods to use the new notification channel and adds a new `callbackShowAzkarNotification` method to handle the periodic showing of notifications. Finally, the commit updates the `pubspec.yaml` file to include the new dependencies.